scripts/scrape_images/requirement.txt
beautifulsoup4==4.12.3
requests
pillow